+++ src/library/entity/RenderEntity.h
@@ -22,7 +22,7 @@ namespace PR
 		virtual bool isLight() const = 0;
 
 		virtual bool isCollidable() const;
-		virtual uint64 collisionCost() const;
+		virtual float collisionCost() const;
 
 		virtual BoundingBox localBoundingBox() const;
 		virtual BoundingBox worldBoundingBox() const;
