+++ src/library/geometry/IMesh.h
@@ -15,7 +15,7 @@ namespace PR
 		virtual bool isLight() const = 0;
 		virtual BoundingBox boundingBox() const = 0;
 		virtual bool checkCollision(const Ray& ray, FacePoint& collisionPoint, float& t) = 0;
-		virtual uint64 collisionCost() const = 0;
+		virtual float collisionCost() const = 0;
 		virtual FacePoint getRandomFacePoint(Sampler& sampler, uint32 sample) const = 0;
 		virtual void replaceMaterial(Material* mat) = 0;
 	};
